+++ samples/Net8AppsettingsJsonSample/LogProperties.cs
@@ -1,0 +1,12 @@
+namespace Net8AppsettingsJsonSample;
+
+/// <summary>
+/// A class to hold the log context properties.
+/// </summary>
+public static class LogProperties
+{
+    /// <summary>
+    /// The correlation id property name.
+    /// </summary>
+    public const string CORRELATION_ID = "CorrelationId";
+}